Move querier to separate package
